+++ src/components/FormComponents/OperationPicker.js
@@ -1,6 +1,6 @@
-import FormComponentGenerator from './FormComponentGenerator';
+import PickerGenerator from './PickerGenerator';
 
-export default FormComponentGenerator({
+export default PickerGenerator({
   defaultLabel: 'Operation ID',
   fields: [{
     type: 'text',
