+++ src/components/TxBuilderAttributes.js
@@ -1,11 +1,11 @@
 import React from 'react';
-import {PubKeyParameter} from './FormComponents/PubKeyParameter';
+import {PubKeyPicker} from './FormComponents/PubKeyPicker';
 
 export default class TxBuilderAttributes extends React.Component {
   render() {
     return <div className="TransactionAttributes">
       <div className="TransactionOp__config TransactionOpConfig optionsTable">
-        <PubKeyParameter param='source_account' />
+        <PubKeyPicker param='source_account' />
       </div>
     </div>
   }
